chore: 전역 Decorator에 GlobalStyle, QueryClient, 라우터 컨텍스트 설정 추가
